feat : cam api 구현

+++ backend/src/main/java/ssafy/age/backend/cam/persistence/CamRepository.java
@@ -2,6 +2,6 @@ package ssafy.age.backend.cam.persistence;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CamRepository extends JpaRepository<Cam, Integer> {
+public interface CamRepository extends JpaRepository<Cam, Long> {
 
 }
